refactor ExpressionCache
